reworked the --help and --version handling cases
Previously it was doing a throw..catch pattern.  This was cumbersome
to handle when I went to go use it in practice.

I'm changing to just raise custom exceptions.  This makes implementation
easier as the user just wraps usage with a begin..rescue and then
handles each exception appropriately.

The option help msg is raised as the HelpExit's message so the user
can access it.