added a `data` reader for accessing the full parsed data set
So in further testing and practice, I found it helpful to have the
`args` available separate from the `opts` for inspection and validation.
However it is still very useful to have a reader that combines them
in a way that is convenient for passing to a handler of some sort.

I went back to having `args` be the arguments and added a `data`
reader that combines the arguments with the trailing opts hash.