removed `:default` handling, switched to `:value` handling
This allows you to pass both and actual value or a class to the
`:value` setting.  If now value is given, `nil` will be used.

I switched to this because in testing and practical use, there were
some scenarios where having 'nothing' as a parsed opt value is
desirable.  Don't want to block this for developers.

I'm also removing references to 'micro-optparse'.  The implementation
and behavior and goals are now quite different.