always use the given option name as the option name
Even if it is a symbol or some weird value. The user needs to
expect to find their named option in the options hash using the
exact name value they gave.

Previously, this was converting all option names to downcased
strings. This was not documented in the README which used symbol
option names and then showed retreiving those option values using
the symbol name. This gets the implementation to matcht he README.
Plus this follows the principle of least surprise I think.